fix(vg_lite): fix CI warning

+++ src/others/vg_lite_tvg/vg_lite_tvg.cpp
@@ -2593,9 +2593,13 @@ static Result picture_load(vg_lite_ctx * ctx, std::unique_ptr<Picture> & picture
     vg_lite_uint32_t * image_buffer;
     LV_ASSERT(VG_LITE_IS_ALIGNED(source->memory, LV_VG_LITE_THORVG_BUF_ADDR_ALIGN));
 
+    /**
+     * Since ThorVG's picture->load does not support stride,
+     * reconversion is required when the stride and width do not match.
+     */
     if(source->format == VG_LITE_BGRA8888
        && source->image_mode == VG_LITE_NORMAL_IMAGE_MODE
-       && source->stride == source->width * sizeof(vg_lite_uint32_t)) {
+       && (size_t)source->stride == (size_t)(source->width * sizeof(vg_lite_uint32_t))) {
         image_buffer = (vg_lite_uint32_t *)source->memory;
     }
     else {
